requirements.txt
python-dotenv
selenium
beautifulsoup4
requests
wget
schedule